requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt>=5.0.0,<6.0
django-cors-headers>=3.13.0,<4.0